+++ changelog.md
@@ -15,6 +15,7 @@
 - twm focus side monitor when theres no more nodes at the specified side in the current workspace.
 - twm allow move windows between monitors via shortcuts.
 - add variables system on extended yaml syntax.
+- new bundled plugins/layouts for the tiling window manager.
 
 ### refactor